Fluid gate M3+M6: hoist core fluid interfaces to :common over IFluidStack + NeoForge cap bridge
The fluid gate master slice. Moves the 4 core fluid interfaces (IExtendedFluidTank,
IMekanismFluidHandler, ISidedFluidHandler, IExtendedFluidHandler) + Extended-
FluidHandlerUtils to :common, retyped over the loader-neutral IFluidStack (drops
the NeoForge IFluidTank/IFluidHandler supers + the @Deprecated fill/drain/
FluidAction overrides; stack-amount ints widened to long, capacity stays int;
mirrors the chemical gate IChemicalTank/ChemicalUtils shapes, method names kept
to avoid ~220 call-site renames).

NeoForge capability exposure PRESERVED byte-identically (M6 bridge): Capabilities.
FLUID stays BlockCapability<IFluidHandler>; the only objects exposed as it --
ProxyFluidHandler (block/entity) + ComponentBackedFluidHandler (item) -- are
rewritten as pure NeoForge IFluidHandler adapters delegating to the :common
Mekanism handlers via NeoFluidStack.wrap/unwrap (fill/drain replicate the old
@Deprecated defaults incl. the item INTERNAL automation type). Impl tank classes
(BasicFluidTank/VariableCapacity/ComponentBacked/MergedTank/etc.), AttachedFluids,
recipes, CrT/JEI/ProjectE stay :neoforge. AttachedFluids bridged at Capability-
TileEntity.apply/collectFluidTanks; recipe-api leaves bridge via new neoforge
api NeoFluidStackBridge. ~46 caller files migrated (networks, multiblock data,
slots, items, GUIs, pump/plenisher, computer); int<->long via saturating casts.
NeoFluidStack gains public static wrap/unwrap.

CapabilityHandlerManager/BasicSidedCapabilityResolver SIDED_HANDLER bound
relaxed to unconstrained (the fluid proxy adapter spans the mek<->NeoForge
boundary; chemical/energy/item/heat still pass compatible types).

Executed via a thorough subagent + reviewed; all 3 modules + all 8 neoforge
source sets compile green (fixed 1 gameTest straggler). Validated :fabric:
runServer all 11 self-tests PASS (fluid mutators=true). NeoForge behavior
preserved by construction (untestable here) -- user must in-game verify fluid
I/O on tanks/pipes/machine-GUIs/multiblocks + cross-mod fluid interop +
existing-world save compat (per docs/FLUID_GATE_PLAN.md M6 risk list).

+++ neoforge/src/generators/java/mekanism/generators/common/content/fusion/FusionReactorMultiblockData.java
@@ -340,7 +340,7 @@ public class FusionReactorMultiblockData extends MultiblockData {
         double caseWaterHeat = MekanismGeneratorsConfig.generators.fusionWaterHeatingRatio.get() * (lastCaseTemperature - biomeAmbientTemp);
         if (Math.abs(caseWaterHeat) > HeatAPI.EPSILON) {
             int waterToVaporize = (int) (HeatUtils.getSteamEnergyEfficiency() * caseWaterHeat / HeatUtils.getWaterThermalEnthalpy());
-            waterToVaporize = Math.min(waterToVaporize, Math.min(waterTank.getFluidAmount(), Ints.saturatedCast(steamTank.getNeeded())));
+            waterToVaporize = (int) Math.min(waterToVaporize, Math.min(waterTank.getFluidAmount(), Ints.saturatedCast(steamTank.getNeeded())));
             if (waterToVaporize > 0) {
                 MekanismUtils.logMismatchedStackSize(waterTank.shrinkStack(waterToVaporize, Action.EXECUTE), waterToVaporize);
                 steamTank.insert(MekanismChemicals.STEAM.asStack(waterToVaporize), Action.EXECUTE, AutomationType.INTERNAL);
